requirements.txt
streamlit
ultralytics
Pillow
pdf2image
opencv-python-headless
PyMuPDF
numpy
gdown
easyocr
langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters<0.3.0
langchain-core<0.3.0
langchain-google-genai<2.0.0
faiss-cpu
sentence-transformers
paddleocr
paddlepaddle
langchain-huggingface<0.1.0
